+++ editor/debugger/editor_debugger_tree.cpp
@@ -98,8 +98,7 @@ void EditorDebuggerTree::_scene_tree_selected() {
 		item->select(0);
 	}
 
-	// `TypedArray<uint64_t>` declares `Variant::INT` elements, so ids are carried nominally.
-	const int64_t id = item->get_metadata(0);
+	const uint64_t id = item->get_metadata(0);
 	inspected_object_ids.append(id);
 
 	if (!notify_selection_queued) {
@@ -113,7 +112,7 @@ void EditorDebuggerTree::_scene_tree_selection_changed(TreeItem *p_item, int p_c
 		return;
 	}
 
-	const int64_t id = p_item->get_metadata(0);
+	const uint64_t id = p_item->get_metadata(0);
 	if (p_selected) {
 		if (inspected_object_ids.size() == (int)EDITOR_GET("debugger/max_node_selection")) {
 			selection_surpassed_limit = true;
@@ -263,7 +262,9 @@ void EditorDebuggerTree::update_scene_tree(const SceneDebuggerTree *p_tree, int 
 		if (icon.is_valid()) {
 			item->set_icon(0, icon);
 		}
-		item->set_metadata(0, node.id);
+		// The selection array validates against the unsigned carrier, so the metadata this is read
+		// back into carries the id there too.
+		item->set_metadata(0, uint64_t(node.id));
 
 		String current_path;
 		if (parent) {
@@ -278,8 +279,8 @@ void EditorDebuggerTree::update_scene_tree(const SceneDebuggerTree *p_tree, int 
 
 		// Select previously selected nodes.
 		if (debugger_id == p_debugger) { // Can use remote id.
-			if (inspected_object_ids.has(node.id)) {
-				ids_present.append(node.id);
+			if (inspected_object_ids.has(uint64_t(node.id))) {
+				ids_present.append(uint64_t(node.id));
 
 				if (selection_uncollapse_all) {
 					selection_uncollapse_all = false;
